requirements.txt
Django>=4.2
djangorestframework>=3.15
django-cors-headers>=4.3
Pillow>=10.0
python-decouple>=3.8
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
